# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "TrelloDash API",
    "description": "API for TrelloDash",
    "version": "0.5.0"
  },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/api/setboard": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/title": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/counts/opentasks": {
      "get": {
        "description": "Returns the amount of all open tasks on the board",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/counts/closedtasks": {
      "get": {
        "description": "Returns the amount of all closed tasks on the board",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/counts/alltasks": {
      "get": {
        "description": "Returns the amount of all tasks on the board",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/cards": {
      "get": {
        "description": "Returns all cards on the board",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/cards/{cardid}": {
      "get": {
        "description": "Returns a specific card on the board",
        "parameters": [
          {
            "name": "cardid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/cards/{cardid}/duration": {
      "get": {
        "description": "Returns the duration of a specific card on the board in minutes",
        "parameters": [
          {
            "name": "cardid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/cards/{cardid}/due": {
      "get": {
        "description": "Returns the due time of a specific card on the board in UNIX-Timecode",
        "parameters": [
          {
            "name": "cardid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/cards/{cardid}/overdue": {
      "get": {
        "description": "Returns the overtime of a specific card on the board in UNIX-Timecode",
        "parameters": [
          {
            "name": "cardid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/criticaltasks": {
      "get": {
        "description": "Returns a list of all critical tasks on the board",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/lists": {
      "get": {
        "description": "Returns all lists on the board",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/lists/{listid}": {
      "get": {
        "description": "Returns a specific list on the board",
        "parameters": [
          {
            "name": "listid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/lists/{listid}/name": {
      "get": {
        "description": "Returns the name of a specific list on the board",
        "parameters": [
          {
            "name": "listid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/lists/{listid}/countall": {
      "get": {
        "description": "Returns the amount of all cards on a specific list on the board",
        "parameters": [
          {
            "name": "listid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/lists/{listid}/countclosed": {
      "get": {
        "description": "Returns the amount of all closed cards on a specific list on the board",
        "parameters": [
          {
            "name": "listid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/lists/{listid}/members": {
      "get": {
        "description": "Returns an array of all members on a specific list on the board",
        "parameters": [
          {
            "name": "listid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/lists/{listid}/overtimed": {
      "get": {
        "description": "Returns an array of all overtimed cards on a specific list on the board",
        "parameters": [
          {
            "name": "listid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/lists/{listid}/anyovertimed": {
      "get": {
        "description": "Returns true if any card on a specific list on the board is overtimed",
        "parameters": [
          {
            "name": "listid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/lists/{listid}/sv": {
      "get": {
        "description": "Returns the summed Schedule Variance of a specific list on the board in minutes",
        "parameters": [
          {
            "name": "listid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/board/sv": {
      "get": {
        "description": "Returns the summed Schedule Variance of the whole board in minutes",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/board/ru": {
      "get": {
        "description": "Returns the summed Resource Utilization of the whole board in decimal notation",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/board/spi": {
      "get": {
        "description": "Returns the summed Schedule Performance Index of the whole board in decimal notation",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/members": {
      "get": {
        "description": "Returns all members on the board",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/members/{memberid}": {
      "get": {
        "description": "Returns a specific member on the board",
        "parameters": [
          {
            "name": "memberid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/members/{memberid}/availabletime": {
      "get": {
        "description": "Returns the available time of a specific member on the board",
        "parameters": [
          {
            "name": "memberid",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/timecards": {
      "get": {
        "description": "Returns an array of all timecards on the board",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/timeplanned": {
      "get": {
        "description": "Returns the planned time (PV) of the whole board in minutes",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}